api/requirements.txt
Django==3.1.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
drf-yasg==1.20.0
django-filter
PyJWT==2.0.1